Тесты для 2.1
Unit тесты сделал Claude Opus 5

+++ module2/2/2.1/phoneBook.h
@@ -47,8 +47,8 @@ typedef struct Contact
     char surname[MAX_SIZE];
     char patronymic[MAX_SIZE];
        
-    size_t countNumbers;
-    size_t countSocial;
+    int countNumbers;
+    int countSocial;
     
     Numbers phonenumber[MAX_COUNT];
     Social social[MAX_COUNT];
